agent/requirements.txt
livekit-agents>=0.12.18,<1.0.0
livekit-plugins-groq>=0.1.2,<1.0.0
livekit-plugins-silero>=0.7.5,<1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
# Sarvam AI plugin (no extra deps — uses httpx)
